% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{a2_load}
\alias{a2_load}
\title{a2Model Load}
\usage{
a2_load(path, sc = NULL)
}
\arguments{
\item{path}{directory path to write a2modeler object to}

\item{sc}{optional spark connection argument. Required to load an a2modeler
object with a spark_model final model type}
}
\description{
Load a2 modeler object into memory to apply on new datasets
}
