% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_cross_validation_samples}
\alias{add_cross_validation_samples}
\alias{add_cross_validation_samples.numeric}
\alias{add_cross_validation_samples.data.frame}
\alias{add_cross_validation_samples.tbl_spark}
\alias{add_cross_validation_samples.modeler}
\alias{add_cross_validation_samples.forecaster}
\title{Add Cross Validation Samples function}
\usage{
add_cross_validation_samples(x, folds, test_holdout_prct, seed)

\method{add_cross_validation_samples}{numeric}(x, folds,
  test_holdout_prct = NULL, seed = NULL)

\method{add_cross_validation_samples}{data.frame}(x, folds,
  test_holdout_prct = NULL, seed = NULL)

\method{add_cross_validation_samples}{tbl_spark}(x, folds,
  test_holdout_prct = NULL, seed = NULL)

\method{add_cross_validation_samples}{modeler}(x, folds,
  test_holdout_prct = NULL, seed = NULL)

\method{add_cross_validation_samples}{forecaster}(x, folds,
  test_holdout_prct = NULL, seed = NULL)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create samples from}

\item{folds}{number of cross validation folds}

\item{test_holdout_prct}{amount of data to holdout out for test sample}

\item{seed}{optional random seed generator}
}
\value{
either an updated modeler object if provided otherwise a samples
  object with resamples
}
\description{
Function creates a new cross_validation sample object based on configuration.
Each index is fold of the dataset
}
\details{
The function creates train and validation pairs using all of the dataset rows
}
\examples{

# Data.frame example
add_cross_validation_samples(mtcars, folds = 4)
}
