% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_default_samples}
\alias{add_default_samples}
\alias{add_default_samples.numeric}
\alias{add_default_samples.data.frame}
\alias{add_default_samples.tbl_spark}
\alias{add_default_samples.modeler}
\title{Add Default Samples function}
\usage{
add_default_samples(x)

\method{add_default_samples}{numeric}(x)

\method{add_default_samples}{data.frame}(x)

\method{add_default_samples}{tbl_spark}(x)

\method{add_default_samples}{modeler}(x)
}
\arguments{
\item{x}{numeric vector, data.frame, spark dataframe or modeler object to
create samples from}
}
\description{
Function creates a no-frills sampling object. Creates a single train index
only. No validation or test indicies created.
}
\details{
Function used when modeler object created as a place holder for modeler
object. Default sampling can be used for simple out of the box models.
}
