% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_model}
\alias{add_model}
\title{Add Model to Modeler Object function}
\usage{
add_model(obj, pipe = NULL, method, param_map = list(), ..., desc = NULL,
  uid = sparklyr::random_string("model"))
}
\arguments{
\item{obj}{modeler object}

\item{method}{string input of model method}

\item{param_map}{list of parameters values to be used in model tuning}

\item{desc}{optional model description}

\item{uid}{model uid}
}
\value{
modeler object with model added
}
\description{
Function to add model to modeler object. More than one model can be added to
a modeler object.
}
\details{
Function creates a new model object from inputs and then appends to modeler
models list
}
