% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{add_model_grid}
\alias{add_model_grid}
\title{Add Model Grid to Modeler Object function}
\usage{
add_model_grid(obj, pipe = NULL, method, ..., desc = NULL, path = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{pipe}{pipeline object. default is empty pipeline which applies no
data transformations}

\item{method}{string input of model method}

\item{...}{additional arguments to pass to model method}

\item{desc}{optional model description}

\item{path}{optional file path to save model}
}
\value{
modeler object with model added
}
\description{
Function to add grid of multiple models to modeler object. Vectors of
parameter values can be provided to this function and grid of models added to
modeler object
}
\details{
Function creates a new model object from inputs and then appends to modeler
models list
}
