% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{add_time_slice_samples}
\alias{add_time_slice_samples}
\alias{add_time_slice_samples.numeric}
\alias{add_time_slice_samples.data.frame}
\alias{add_time_slice_samples.tbl_spark}
\alias{add_time_slice_samples.forecaster}
\title{Add Time Slice function}
\usage{
add_time_slice_samples(x, width, horizon, skip = 0, fixed_width = TRUE)

\method{add_time_slice_samples}{numeric}(x, width, horizon, skip = 0,
  fixed_width = TRUE)

\method{add_time_slice_samples}{data.frame}(x, width, horizon, skip = 0,
  fixed_width = TRUE)

\method{add_time_slice_samples}{tbl_spark}(x, width, horizon, skip = 0,
  fixed_width = TRUE)

\method{add_time_slice_samples}{forecaster}(x, width, horizon, skip = 0,
  fixed_width = TRUE)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create samples from}

\item{width}{number of training sample periods}

\item{horizon}{number of test sample periods}

\item{skip}{number of samples skipped. used to thin sample pairs}

\item{fixed_width}{logical option to keep all training widths the same.
default is TRUE}
}
\value{
either an updated modeler object if provided otherwise a samples
  object with resamples
}
\description{
Function creates a new samples object with time slice resamples based on
configuration. Each index is time slice sample of the dataset
}
\details{
The function creates train and validation pairs
}
\examples{

# Basic Time Slice example
add_time_slice_samples(1:20, width = 15, horizon = 1)

# Time Slice with Skip
 add_time_slice_samples(1:20, width = 15, horizon = 1, skip = 1)
}
