% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{apply_model}
\alias{apply_model}
\title{Apply Model Generic Fuction}
\usage{
apply_model(mobj, ...)
}
\arguments{
\item{mobj}{model object}

\item{...}{additional arguments to pass to fit function}
}
\value{
updated model object
}
\description{
Function applies model to make predictions for all samples
}
\details{
Fitted values extracted for training data and predictions made for validation
or test samples. Predictions stored with sub-model
}
