% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{as_data_frame.forecast}
\alias{as_data_frame.forecast}
\title{Convert Forecast object to data.frame}
\usage{
as_data_frame.forecast(mobj)
}
\arguments{
\item{mobj}{forecast model object}
}
\description{
Convert Forecast object to data.frame
}
