% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{auc}
\alias{auc}
\alias{auc.tbl_spark}
\alias{auc.spark_connection}
\title{Generic auc function}
\usage{
auc(x, predicted, actual)

\method{auc}{tbl_spark}(x, predicted, actual)

\method{auc}{spark_connection}(x, predicted = "predicted", actual = "label")
}
\arguments{
\item{x}{dataframe object}

\item{predicted}{column name of predicted values}

\item{actual}{column name of actual values}
}
\description{
Generic auc function
}
