% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{clean_pipes}
\alias{clean_pipes}
\title{Flush Pipelines function}
\usage{
clean_pipes(obj, uids = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{uids}{one or more model ids. default is null which cleans all pipes}
}
\value{
updated modeler object
}
\description{
Function to clean pipelines for one or more models in modeler object
}
