% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{cross_validation}
\alias{cross_validation}
\alias{cross_validation.numeric}
\alias{cross_validation.data.frame}
\alias{cross_validation.tbl_spark}
\title{Cross Validation Sample function}
\usage{
cross_validation(x, folds)

\method{cross_validation}{numeric}(x, folds)

\method{cross_validation}{data.frame}(x, folds)

\method{cross_validation}{tbl_spark}(x, folds)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create samples from}

\item{folds}{number of cross validation folds}
}
\description{
Function creates a cross validation sample index from a dataset. Index can be used to
create data samples
}
