% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{deploy}
\alias{deploy}
\title{Deploy Modeler Object}
\usage{
deploy(obj, path, ...)
}
\arguments{
\item{obj}{modeler object to save}

\item{path}{directory path to write a2modeler object to}

\item{...}{additional arguments to pass to \link[base]{saveRDS} function}
}
\description{
Save Modeler object to file that can be restored later. Requires a final
model to be set. Deploy uses \code{\link{saveRDS}} function to serialize the
R object. Deploy uses \code{\link{sparklyr::ml_save}} to serialize final
model fit.
}
\details{
Deploy function creates a new directory at path location and writes a2modeler
object and final model within
}
