% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-model-clustering.R
\name{evaluate.spark_ml_clustering}
\alias{evaluate.spark_ml_clustering}
\title{Evaluate Method for Spark-Model Clustering Object}
\usage{
evaluate.spark_ml_clustering(mobj, measure)
}
\description{
Evaluate Method for Spark-Model Clustering Object
}
