% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{evaluate_models}
\alias{evaluate_models}
\alias{evaluate_models.modeler}
\title{Evaluate Models Generic Function}
\usage{
evaluate_models(obj, ids, ...)

\method{evaluate_models}{modeler}(obj, ids = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{ids}{optional input for model id. default is NULL and all trained
models evaluated}
}
\value{
updated modeler object
}
\description{
Evaluate the accuracy of all trained models.
}
\details{
Function applies the measure function associated with modeler object to all
model sample predictions.
}
