% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{execute_pipelines}
\alias{execute_pipelines}
\alias{execute_pipelines.modeler}
\title{Execute Pipelines Generic Function}
\usage{
execute_pipelines(obj, ...)

\method{execute_pipelines}{modeler}(obj, uids = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{...}{additional arguments to pass to execute function}
}
\value{
updated modeler object
}
\description{
Function executes all pipelines added to modeler object
}
\details{
Only executes previous un-executed pipelines
}
