% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{extend}
\alias{extend}
\alias{extend.index}
\alias{extend.time_index}
\title{Extend an Index}
\usage{
extend(obj, length_out)

\method{extend}{index}(obj, length_out)

\method{extend}{time_index}(obj, length_out)
}
\arguments{
\item{obj}{index object}

\item{length_out}{number of periods to extend index by}
}
\description{
Function to extend out an index by specified length
}
