% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{f1}
\alias{f1}
\alias{f1.tbl_spark}
\alias{f1.spark_connection}
\title{Generic f1 function}
\usage{
f1(x, predicted, actual)

\method{f1}{tbl_spark}(x, predicted, actual)

\method{f1}{spark_connection}(x, predicted = "prediction", actual = "label")
}
\arguments{
\item{x}{dataframe object}

\item{predicted}{column name of predicted values}

\item{actual}{column name of actual values}
}
\description{
Generic f1 function
}
