% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_model}
\alias{fit_model}
\title{Fit Model Generic Function}
\usage{
fit_model(mobj, ...)
}
\arguments{
\item{mobj}{model object}

\item{...}{additional arguments to pass to fit model function}
}
\value{
updated model object
}
\description{
Functions execues model method on its pipeline output for given param grid
and sampling indicies
}
\details{
Sub model fits are stored in sub-model list. sub-models given unique id. Each
sub-model is fit on each training sample provided
}
