% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/spark-ml.R
\name{fitted.forecast_model}
\alias{fitted.forecast_model}
\alias{fitted.spark_ml}
\alias{fitted.ml_decision_tree_classification_model}
\title{Forecast Model Fitted Method}
\usage{
\method{fitted}{forecast_model}(mobj)

\method{fitted}{spark_ml}(mobj)

\method{fitted}{ml_decision_tree_classification_model}(mobj)
}
\arguments{
\item{mobj}{model object}
}
\description{
Extracts fitted values from train step

Extracts fitted values from train step

Fitted Method for Spark-ML Decision Tree Classification Model
}
