% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_best_model}
\alias{get_best_model}
\title{Get the Best Performing Model}
\usage{
get_best_model(obj)
}
\arguments{
\item{obj}{modeler object}
}
\value{
returns best model
}
\description{
Function to select the best performing model based on modeler measure
}
