% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_coefs}
\alias{get_coefs}
\title{Return the Model Coefficients}
\usage{
get_coefs(mobj, ...)
}
\arguments{
\item{mobj}{model object}

\item{...}{additional arguments to pass through}
}
\description{
Return the Model Coefficients
}
