% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_evalutions}
\alias{get_evalutions}
\title{Get Model Evaluation Summary}
\usage{
get_evalutions(obj)
}
\arguments{
\item{obj}{modeler object}
}
\description{
Returns data.frame with measure calculated on the validation samples
}
