% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/model.R
\name{get_forecasts.forecast}
\alias{get_forecasts.forecast}
\alias{get_forecasts}
\title{Get Forecasts from forecast object}
\usage{
\method{get_forecasts}{forecast}(mobj)

get_forecasts(mobj, ...)
}
\arguments{
\item{mobj}{model object}

\item{...}{additional arguments to pass through}
}
\value{
data.frame with forecats from forecast model
}
\description{
Get Forecasts from forecast object

Get Model Forecasts
}
