% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{get_measure}
\alias{get_measure}
\title{Get Modeler Measure}
\usage{
get_measure(obj)
}
\arguments{
\item{obj}{modeler object}
}
\value{
measure object
}
\description{
Getter function to extract measure from modeler object
}
