% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_models_status}
\alias{get_models_status}
\title{Get Models Status}
\usage{
get_models_status(obj, uid = NULL)
}
\arguments{
\item{obj}{modeler object}

\item{uid}{one or more model uid characters}
}
\description{
Function to return modeler's model statuses
}
