% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_performance}
\alias{get_performance}
\alias{get_performance}
\title{Get Model Performance}
\usage{
get_performance(obj)

get_performance(obj)
}
\arguments{
\item{obj}{object to extract target from}
}
\description{
Returns data.frame with model performance
}
