% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_schema}
\alias{get_schema}
\alias{get_schema.data.frame}
\alias{get_schema.tbl_spark}
\title{Get Dataset Schema}
\usage{
get_schema(df)

\method{get_schema}{data.frame}(df)

\method{get_schema}{tbl_spark}(df)
}
\arguments{
\item{df}{dataset}
}
\value{
named list of column types
}
\description{
Returns named list of column types. Names refer to column names
}
\examples{
get_schema(mtcars)
}
