% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/modeler.R
\name{get_target.model}
\alias{get_target.model}
\alias{get_target}
\alias{get_target.modeler}
\title{Get Target Data function}
\usage{
\method{get_target}{model}(obj)

get_target(obj)

\method{get_target}{modeler}(obj)
}
\arguments{
\item{obj}{object to extract target from}
}
\description{
Returns modeler target data
}
