% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_target_df}
\alias{get_target_df}
\alias{get_target_df.data.frame}
\title{Get Target Dataframe Generic Method}
\usage{
get_target_df(df, target)

\method{get_target_df}{data.frame}(df, target)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{target}{target variable for supervised use cases such as forecaster,
regressor, and classifier}
}
\description{
Returns modeler target data
}
