% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-model.R, R/model.R, R/spark-ml.R,
%   R/spark-model.R
\name{get_variable_importance.forecast_model}
\alias{get_variable_importance.forecast_model}
\alias{get_variable_importance}
\alias{get_variable_importance.default}
\alias{get_variable_importance.ml_linear_regression_model}
\alias{get_variable_importance.ml_logistic_regression_model}
\alias{get_variable_importance.ml_decision_tree_regression_model}
\alias{get_variable_importance.ml_decision_tree_classification_model}
\alias{get_variable_importance.ml_random_forest_regression_model}
\alias{get_variable_importance.ml_random_forest_classification_model}
\alias{get_variable_importance.ml_gbt_regression_model}
\alias{get_variable_importance.ml_gbt_classification_model}
\alias{get_variable_importance.spark_model}
\title{Get Model Variable Importance Metrics}
\usage{
\method{get_variable_importance}{forecast_model}(mobj)

get_variable_importance(mobj, ...)

\method{get_variable_importance}{default}(mobj, ...)

\method{get_variable_importance}{ml_linear_regression_model}(mobj,
  features)

\method{get_variable_importance}{ml_logistic_regression_model}(mobj,
  features)

\method{get_variable_importance}{ml_decision_tree_regression_model}(mobj,
  features)


  \method{get_variable_importance}{ml_decision_tree_classification_model}(mobj,
  features)

\method{get_variable_importance}{ml_random_forest_regression_model}(mobj,
  features)


  \method{get_variable_importance}{ml_random_forest_classification_model}(mobj,
  features)

\method{get_variable_importance}{ml_gbt_regression_model}(mobj, features)

\method{get_variable_importance}{ml_gbt_classification_model}(mobj,
  features)

\method{get_variable_importance}{spark_model}(mobj)
}
\arguments{
\item{mobj}{model object}
}
\description{
Returns variable importance from tree based methods, coefficient summaries
from linear based methods, and NULL for methods without inferential metrics
}
