% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{holdout}
\alias{holdout}
\alias{holdout.numeric}
\alias{holdout.data.frame}
\alias{holdout.tbl_spark}
\title{Holdout Sampling function}
\usage{
holdout(x, split)

\method{holdout}{numeric}(x, split)

\method{holdout}{data.frame}(x, split)

\method{holdout}{tbl_spark}(x, split)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create samples from}

\item{split}{numeric input for percentage of rows in head index}
}
\description{
Function creates two row indicies from a dataset based on split parameter.
Indicies used to create data samples
}
