% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{mape}
\alias{mape}
\alias{mape.data.frame}
\alias{mape.tbl_spark}
\title{Generic mape function}
\usage{
mape(x, predicted, actual)

\method{mape}{data.frame}(x, predicted, actual)

\method{mape}{tbl_spark}(x, predicted, actual)
}
\arguments{
\item{x}{dataframe object}

\item{predicted}{column name of predicted values}

\item{actual}{column name of actual values}
}
\description{
Generic mape function
}
