% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{match_measure_method}
\alias{match_measure_method}
\title{Match Modeler Measurement Method}
\usage{
match_measure_method(obj)
}
\arguments{
\item{obj}{modeler object}
}
\value{
measure method function
}
\description{
Function to match the internal modeler measure method
}
