% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measure}
\alias{measure}
\title{Measure Class Constructer}
\usage{
measure(id, method, method_args = list(), properties, name = id, minimize,
  best = NULL, worst = NULL, note = "")
}
\arguments{
\item{id}{name of measure ex - mse}

\item{method}{name of measure method. method function differ based on class
of input. Desigend to be accomdate spark dataframes}

\item{method_args}{([list]) additional arguments to pass to method}

\item{properties}{measure's applicaple modeler classes}

\item{name}{full name of measure. ex - mean squared error}

\item{minimize}{logical flag that minimizing the measure is optimal}

\item{best}{best possible value for measure}

\item{worst}{worst possible value for measure}

\item{note}{additional notes for measure}
}
\description{
Measure Class Constructer
}
\note{
measure class designed after similiar named class in mlr package:
  \url{https://github.com/mlr-org/mlr/blob/master/R/Measure.R}
}
