% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{RMSE}
\alias{RMSE}
\alias{MAPE}
\alias{Silhouette}
\alias{AUC}
\title{Root Mean Squared Error Measure Object}
\format{An object of class \code{measure} of length 9.}
\usage{
RMSE

MAPE

Silhouette

AUC
}
\description{
Root Mean Squared Error Measure Object

rmse.tbl_spark <- function(x, predicted, actual){
  checkmate::assert_choice(predicted, colnames(x))
  checkmate::assert_choice(actual, colnames(x))
  x %>%
    dplyr::summarise_at(predicted, funs(rmse = sqrt(mean((. - !!rlang::sym(actual))^2)))) %>%
    dplyr::collect()
}
Mean Absolute Percentage Error Measure Object

Silhouette Measure Object

Area Under the Curve Measure Object
}
\keyword{datasets}
