% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Model Class Helper}
\usage{
model(pipe, target, method, param_grid, desc = NULL, uid = NULL)
}
\arguments{
\item{target}{column name of target variable. string input}

\item{method}{string input of model method}

\item{param_grid}{parameter grid for model tuning}

\item{desc}{optional model description}

\item{uid}{model uid}

\item{pipe_uid}{pipeline uid string}
}
\description{
Function to create a model object.
}
\details{
Requires a pipeline object input, and a valid model method. Any method
package dependencies need to be loaded prior to model call
}
