% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{new_classifier}
\alias{new_classifier}
\alias{classifier}
\title{Classifier Constructer Function}
\usage{
new_classifier(df, target, name = NULL, id = NULL, version = NULL,
  desc = NULL, scientist = NULL, save_fits = TRUE, dir = NULL, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{target}{target variable for supervised use cases such as forecaster,
regressor, and classifier}

\item{name}{optional name for use case. Default is to append target with type}

\item{id}{optional id string. A random id will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{save_fits}{logical flag to save model fits in modeler object. note -
does not apply to final model}

\item{dir}{optional directory path for saving reports, data and models}

\item{...}{any addtional parameters}
}
\description{
Creates a classifier object which inherits from modeler class
}
\seealso{
Other use cases: \code{\link{modeler}},
  \code{\link{new_forecaster}},
  \code{\link{new_regressor}}, \code{\link{new_segmenter}}
}
