% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{new_model}
\alias{new_model}
\title{Model Class Constructer}
\usage{
new_model(pipe, target, method, method_args, param_map, desc, uid, status,
  created_on)
}
\arguments{
\item{pipe}{pipeline uid string}

\item{target}{column name of target variable. string input}

\item{method}{string input of model method}

\item{method_args}{required method arguments not in parameter grid}

\item{param_map}{list of parameter values for model tuning}

\item{desc}{optional model description}

\item{uid}{model uid}
}
\description{
Model Class Constructer
}
