% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{new_modeler}
\alias{new_modeler}
\title{New Modeler Object Constructer function}
\usage{
new_modeler(df, schema, target, type, name, id, version, desc, scientist, dir,
  save_fits, samples, models, measure, evaluate, final_model, ...)
}
\description{
New Modeler Object Constructer function
}
