% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{refit}
\alias{refit}
\alias{refit.modeler}
\title{Refit Modeler Object with New Data}
\usage{
refit(obj, df, append = TRUE, ...)

\method{refit}{modeler}(obj, df, append = TRUE, ...)
}
\arguments{
\item{obj}{modeler object}

\item{df}{new dataframe}

\item{append}{logical option to append new dataframe to exisiting modeler
data}

\item{...}{additional arguments to pass to refit function. Not currently
implemented}
}
\value{
updated modeler object with new final model
}
\description{
Refits a modeler's final model to new data. Option to append new data to
existing data and fit on combined data or fit on just new data
}
\details{
Requires newdata have similar schema to existing modeler data
}
