% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{rmse}
\alias{rmse}
\alias{rmse.data.frame}
\alias{rmse.tbl_spark}
\alias{rmse.spark_connection}
\title{Generic rmse function}
\usage{
rmse(x, predicted, actual)

\method{rmse}{data.frame}(x, predicted, actual)

\method{rmse}{tbl_spark}(x, predicted, actual)

\method{rmse}{spark_connection}(x, predicted = "prediction",
  actual = "label")
}
\arguments{
\item{x}{dataframe object}

\item{predicted}{column name of predicted values}

\item{actual}{column name of actual values}
}
\description{
Function that calculates root mean squared error
}
