% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{Salaries of San Fransico Municipal Workers}
\format{A data frame with 148686 rows and 14 variables:
 \describe{
  \item{first_name}{first name}
  \item{last_name}{last name}
  \item{gender_guess}{gender guess based on first name}
  \item{job_title}{job title}
  \item{year}{employment year}
  \item{base_pay}{base salary}
  \item{overtime_pay}{overtime pay}
  \item{other_pay}{other pay}
  \item{total_pay}{base, overtime and other pay total}
  \item{benefits}{total benefit cost}
  \item{total}{total pay plus benefits}
  \item{police}{flag for police based on job title}
  \item{fire}{flag for firefighter based on job title}
  \item{medical}{flag for medical professional based on job title}...
  }}
\usage{
salaries
}
\description{
A dataset containing the San Fransico Municipal workers salaries
}
\keyword{datasets}
