% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmenter.R
\name{segmenter}
\alias{segmenter}
\title{Segmenter Constructer Function}
\usage{
segmenter(df, name = NULL, uid = NULL, version = NULL, desc = NULL,
  scientist = NULL, execution_strategy = "sequential", refit = TRUE,
  save_submodels = TRUE, dir = NULL, seed = 319, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{name}{optional name for use case. Default is to append target with type}

\item{uid}{optional uid string. A random uid will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{execution_strategy}{name of evaluation function to use for resolving a
future. Used for model training execution. default is sequential}

\item{refit}{logical option to refit model to entire training dataset sample.
If grid parameters supplied, the best param set determined by validation
sample performance will be used. Default is true}

\item{save_submodels}{logical flag to save sub model fits. Default is TRUE}

\item{dir}{optional directory path for saving reports, data and models}

\item{seed}{seed number of any random computations}

\item{...}{any addtional parameters}
}
\description{
Creates a segmenter object which inherits from modeler class
}
\seealso{
Other use cases: \code{\link{classifier}},
  \code{\link{forecaster}}, \code{\link{modeler}},
  \code{\link{multiclassifier}}, \code{\link{regressor}}
}
\concept{use cases}
