% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-model.R
\name{set_evaluator}
\alias{set_evaluator}
\title{Set Evaluator Function}
\usage{
set_evaluator(mobj, sc, measure, ...)
}
\arguments{
\item{mobj}{spark_model object}

\item{sc}{spark_connection}

\item{measure}{measure object}
}
\value{
updated modeler object
}
\description{
Function to create new evaluator object used in validator object
}
