% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{set_final_model}
\alias{set_final_model}
\alias{set_final_model.modeler}
\title{Set Final Model Generic}
\usage{
set_final_model(obj, method, uid, reevaluate, refit)

\method{set_final_model}{modeler}(obj, method, uid = NULL,
  reevaluate = TRUE, refit = TRUE)
}
\arguments{
\item{obj}{forecaster object}

\item{method}{selection method. Choices are 'manual' which requires a valid
uid or 'best' method which selects the best model based on the measure}

\item{uid}{optional model uid input required for manual method}

\item{reevaluate}{logical option to re-evaluate the final model. Requires a
test holdout sample}

\item{refit}{logical option to re-fit the final model. Requires a test
holdout sample}
}
\description{
Function to select the final model. See method argument for options
}
