% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{set_measure}
\alias{set_measure}
\title{Set Modeler Method}
\usage{
set_measure(obj, measure)
}
\arguments{
\item{obj}{modeler object}

\item{measure}{measure object}
}
\description{
Function updates or sets measure for modeler object. Requires valid modeler
and measure objects
}
