% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{silhouette}
\alias{silhouette}
\alias{silhouette.tbl_spark}
\title{Generic silhouette function}
\usage{
silhouette(x, predicted)

\method{silhouette}{tbl_spark}(x, predicted = "predicted")
}
\arguments{
\item{x}{dataframe object}

\item{predicted}{column name of predicted values}
}
\description{
Generic silhouette function
}
