% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_df_ts}
\alias{sim_df_ts}
\title{Simulated Time Series Dataset}
\format{A tibble with 200 rows and 2 variables
 \describe{
  \item{index}{integer type with sequential index of records}
  \item{date}{date type. with date of record}
  \item{y}{time series values}
 }}
\usage{
sim_df_ts
}
\description{
A dataset with sequential data created from arima.sim function with random
innovations
}
\keyword{datasets}
