% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R, R/spark-ml.R
\name{summary.forecast_model}
\alias{summary.forecast_model}
\alias{summary.spark_ml}
\title{Summary Method for Spark-ML Object}
\usage{
\method{summary}{forecast_model}(mobj)

\method{summary}{spark_ml}(mobj)
}
\description{
Summary Method for Spark-ML Object
}
