% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{time_slice}
\alias{time_slice}
\alias{time_slice.numeric}
\alias{time_slice.integer}
\alias{time_slice.data.frame}
\alias{time_slice.tbl_spark}
\title{Time Slice Sample function}
\usage{
time_slice(x, width, horizon, skip, fixed_width)

\method{time_slice}{numeric}(x, width, horizon, skip = 0,
  fixed_width = TRUE)

\method{time_slice}{integer}(x, width, horizon, skip = 0,
  fixed_width = TRUE)

\method{time_slice}{data.frame}(x, width, horizon, skip = 0,
  fixed_width = TRUE)

\method{time_slice}{tbl_spark}(x, width, horizon, skip = 0,
  fixed_width = TRUE)
}
\arguments{
\item{x}{numeric vector, dataframe, spark dataframe, or modeler object to
create samples from}

\item{width}{number of training sample periods}

\item{horizon}{number of test sample periods}

\item{skip}{number of samples skipped. used to thin sample pairs}

\item{fixed_width}{logical option to keep all training widths the same.
default is TRUE}
}
\description{
Time Slice Sample function
}
\note{
similiar in design to caret's createTimeSlices function
  \url{https://github.com/topepo/caret/blob/master/pkg/caret/R/createDataPartition.R}
}
