% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/spark-model.R
\name{train_model}
\alias{train_model}
\alias{train_model.spark_model}
\title{Train Model Generic Function}
\usage{
train_model(mobj, ...)

\method{train_model}{spark_model}(mobj, data, measure, samples, save_submodels,
  execution_strategy)
}
\arguments{
\item{mobj}{model object}

\item{...}{additional arguments to pass on}
}
\value{
updated modeler object
}
\description{
Train a model added to a modeler object.
}
\details{
Function fits model based on modeler samples, model pipeline, model method
and param grid.
}
