% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{train_models}
\alias{train_models}
\alias{train_models.modeler}
\title{Train Model Generic Function}
\usage{
train_models(obj, ...)

\method{train_models}{modeler}(obj, ids = NULL)
}
\value{
updated modeler object
}
\description{
Train all models added to a modeler object
}
\details{
Function looks for all untrained models added to a modeler project. Function
fits model based on modeler samples, model method and model method args.
Function saves model fit, all fitted predictions to the training sample and
any validation predictions
}
