% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{deploy}
\alias{deploy}
\title{Deploy Modeler Object}
\usage{
deploy(obj, path, lighten = FALSE, ...)
}
\arguments{
\item{obj}{modeler object to save}

\item{path}{file path to save to. Extension should be .rds}

\item{lighten}{logical flag to lighten the memory footprint of modeler
objects by removing all fit and pipeline output from trained models. Note
the final model fit and pipeline are not impacted}

\item{...}{Arguments passed on to \code{saveRDS}
\describe{
  \item{}{}
}}
}
\description{
Save Modeler object to file that can be restored later. Deploy uses
\code{\link{saveRDS}} function to serialize the R object
}
