% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/spark-ml.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.model}
\alias{evaluate.spark_ml_clustering}
\alias{evaluate.spark_ml_classification}
\alias{evaluate.ml_model_regression}
\title{Evaluate Model Generic}
\usage{
evaluate(mobj, measure)

\method{evaluate}{model}(mobj, measure)

\method{evaluate}{spark_ml_clustering}(mobj, measure)

\method{evaluate}{spark_ml_classification}(mobj, measure)

\method{evaluate}{ml_model_regression}(mobj, measure)
}
\arguments{
\item{mobj}{model object}

\item{measure}{measure object}
}
\value{
returns evaluted model object
}
\description{
Function to evaluate the predictive performance of a model

Evaluate Method for Spark-ML Clustering Object

Evaluate Method for Spark-ML Binary Classification Object

Evaluate Method for Spark-ML Regression Object
}
