% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-model.R, R/model.R,
%   R/spark-model-clustering.R, R/spark-model.R
\name{evaluate_model.forecast_model}
\alias{evaluate_model.forecast_model}
\alias{evaluate_model}
\alias{evaluate_model.spark_ml_clustering}
\alias{evaluate_model.spark_model}
\title{Evaluate Model Generic Function}
\usage{
\method{evaluate_model}{forecast_model}(mobj, data, measure,
  prediction_col = "mean", ...)

evaluate_model(mobj, ...)

\method{evaluate_model}{spark_ml_clustering}(mobj, measure, ...)

\method{evaluate_model}{spark_model}(mobj, data, measure,
  prediction_col = "predicted", ...)
}
\arguments{
\item{mobj}{model object}

\item{data}{data with actual values used to compare predictions against}

\item{measure}{measure object}

\item{prediction_col}{colname of predicted values}

\item{...}{additional arguments to pass to method}
}
\value{
updated modeler object
}
\description{
Evaluate the accuracy of a fitted and applied model
}
\details{
Function applies the measure function associated with modeler object to a
model predictions
}
