% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-ml.R
\name{fitted.spark_model}
\alias{fitted.spark_model}
\title{Fitted Method for Spark Model Object}
\usage{
\method{fitted}{spark_model}(mobj, prediction_col = "predicted")
}
\arguments{
\item{mobj}{spark_model object}
}
\description{
Extracts fitted values from train_model step
}
