% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-model.R, R/spark-ml.R
\name{fitted.forecast_model}
\alias{fitted.forecast_model}
\alias{fitted.spark_model}
\title{Forecast Model Fitted Method}
\usage{
\method{fitted}{forecast_model}(mobj, ...)

\method{fitted}{spark_model}(mobj, prediction_col = "predicted", ...)
}
\arguments{
\item{mobj}{spark_model object}
}
\description{
Extracts fitted values from train step

Extracts fitted values from train_model step
}
