% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-model.R, R/model.R, R/spark-ml.R
\name{get_coefs.forecast_model}
\alias{get_coefs.forecast_model}
\alias{get_coefs}
\alias{get_coefs.ml_model_linear_regression}
\title{Get Coefficients from Forecast Model Object}
\usage{
\method{get_coefs}{forecast_model}(mobj)

get_coefs(mobj)

\method{get_coefs}{ml_model_linear_regression}(mobj)
}
\arguments{
\item{mobj}{model object}
}
\description{
Get Coefficients from Forecast Model Object

Return the Model Coefficients
}
