% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_forecasts}
\alias{get_forecasts}
\title{Get Model Forecasts}
\usage{
get_forecasts(mobj, ...)
}
\arguments{
\item{mobj}{model object}

\item{...}{additional arguments to pass through}
}
\description{
Get Model Forecasts
}
