% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{get_performance}
\alias{get_performance}
\alias{get_performance}
\title{Get Model Performance Generic}
\usage{
get_performance(obj)

get_performance(obj)
}
\description{
Function to extract model performance on samples. Converts output to
data.frame
}
