% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_variable_importance}
\alias{get_variable_importance}
\title{Get Model Variable Importance Metrics}
\usage{
get_variable_importance(mobj)
}
\arguments{
\item{mobj}{model object}
}
\description{
Returns variable importance from tree based models and coefficient summaries
from linear based models
}
