% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index.numeric}
\alias{index.Date}
\alias{index.POSIXct}
\title{Index Creation Function}
\usage{
index(x, unit)

\method{index}{numeric}(x, unit = NULL)

\method{index}{Date}(x, unit = "days")

\method{index}{POSIXct}(x, unit = "hours")
}
\description{
Creates new index object for sequential data management
}
\details{
Used by Forecaster to create forward looking indicies
}
