% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{measures}
\alias{measures}
\alias{RMSE}
\alias{MAPE}
\alias{Silhouette}
\alias{AUC}
\alias{F1}
\title{a2modeler Measures}
\format{An object of class \code{measure} of length 9.}
\usage{
RMSE

MAPE

Silhouette

AUC

F1
}
\description{
Measure objects compatible with a2modeler objects

RMSE - Root Mean Squared Error Measure Object

MAPE - Mean Absolute Percentage Error Measure Object

Silhouette - Silhouette Measure Object

AUC - Area Under the Curve Measure Object

F1 - F1 Measure Object
}
\keyword{datasets}
