% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{new_modeler}
\alias{new_modeler}
\title{New Modeler Object Constructer function}
\usage{
new_modeler(df, schema, target, type, name, uid, version, desc, scientist, dir,
  execution_strategy, refit, save_submodels, samples, pipelines, models,
  measure, performance, final_model, ...)
}
\description{
New Modeler Object Constructer function
}
