% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{new_modeler}
\alias{new_modeler}
\title{New Modeler Object Constructer function}
\usage{
new_modeler(df, schema, target, type, name, uid, version, desc, scientist, dir,
  execution_strategy, refit, save_submodels, samples, pipelines, models,
  measure, performance, final_model, seed, ...)
}
\arguments{
\item{df}{dataframe with target variable and any potential predictors}

\item{target}{target variable for supervised use cases such as forecaster,
regressor, and classifier}

\item{type}{type of use case (forecaster, regressor, classifier, and
segmenter)}

\item{name}{optional name for use case. Default is to append target with type}

\item{uid}{optional uid string. A random uid will be generated by default}

\item{version}{optional version string. will default to '1.0'}

\item{desc}{optional description of the use case.}

\item{scientist}{optional input for scientist building model. defaults to
system user}

\item{dir}{optional directory path for saving reports, data and models}

\item{execution_strategy}{name of evaluation function to use for resolving a
future. Used for model training execution. default is sequential}

\item{refit}{logical option to refit model to entire training dataset sample.
If grid parameters supplied, the best param set determined by validation
sample performance will be used. Default is true}

\item{save_submodels}{logical flag to save sub model fits. Default is TRUE}

\item{seed}{seed number of any random computations}

\item{...}{any addtional parameters}
}
\description{
New Modeler Object Constructer function
}
