% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R, R/forecast.R, R/forecaster.R,
%   R/regressor.R, R/segmenter.R, R/spark-ml.R
\name{predict.classifier}
\alias{predict.classifier}
\alias{predict.forecast_model}
\alias{predict.forecaster}
\alias{predict.regressor}
\alias{predict.segmenter}
\alias{predict.spark_ml_clustering}
\alias{predict.spark_ml_classification}
\alias{predict.ml_model_regression}
\title{Classifier Predict Method}
\usage{
\method{predict}{classifier}(obj, data = NULL, desc = "", ...)

\method{predict}{forecast_model}(obj, periods, data = NULL, level = c(80,
  95))

\method{predict}{forecaster}(obj, periods, data = NULL, level = c(80, 95),
  desc = "")

\method{predict}{regressor}(obj, data = NULL, desc = "", ...)

\method{predict}{segmenter}(obj, data = NULL, desc = "", ...)

\method{predict}{spark_ml_clustering}(mobj, data, ...)

\method{predict}{spark_ml_classification}(mobj, data, ...)

\method{predict}{ml_model_regression}(mobj, data, ...)
}
\description{
Method makes predictions using classifier object's final model

Forecast Prediction Method

Method makes predictions for Forecaster's final model

Method makes predictions using regressor object's final model

Method makes predictions using segmenter object's final model

Predict Method for Spark-ML Clustering Object

Predict Method for Spark-ML Classification Object

Predict Method for Spark-ML Regression Object
}
