% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R, R/forecast-model.R,
%   R/forecaster.R, R/multiclassifier.R, R/regressor.R, R/segmenter.R,
%   R/spark-ml.R, R/spark-model-clustering.R
\name{predict.classifier}
\alias{predict.classifier}
\alias{predict.forecast_model}
\alias{predict.forecaster}
\alias{predict.multiclassifier}
\alias{predict.regressor}
\alias{predict.segmenter}
\alias{predict.spark_model_classification}
\alias{predict.spark_model_regression}
\alias{predict.spark_model_clustering}
\title{Classifier Predict Method}
\usage{
\method{predict}{classifier}(obj, data = NULL, desc = "", ...)

\method{predict}{forecast_model}(mobj, data = NULL, periods,
  level = c(80, 95), ...)

\method{predict}{forecaster}(obj, data = NULL, periods,
  uid = sparklyr::random_string(prefix = "pred"), desc = "", ...)

\method{predict}{multiclassifier}(obj, data = NULL, desc = "", ...)

\method{predict}{regressor}(obj, data = NULL, desc = "", ...)

\method{predict}{segmenter}(obj, data = NULL, desc = "", ...)

\method{predict}{spark_model_classification}(mobj, data,
  prediction_col = "predicted", ...)

\method{predict}{spark_model_regression}(mobj, data,
  prediction_col = "predicted", ...)

\method{predict}{spark_model_clustering}(mobj, data,
  prediction_col = "predicted", ...)
}
\description{
Method makes predictions using classifier object's final model

Method makes predictions for Forecaster's final model

Method makes predictions using multiclassifier object's final model

Method makes predictions using regressor object's final model

Method makes predictions using segmenter object's final model
}
