% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-model.R, R/spark-model.R
\name{summary.forecast_model}
\alias{summary.forecast_model}
\alias{summary.spark_model}
\title{Summary Method for Spark Model Object}
\usage{
\method{summary}{forecast_model}(mobj)

\method{summary}{spark_model}(mobj)
}
\description{
Summary Method for Spark Model Object
}
