% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-model.R, R/model.R,
%   R/spark-model-clustering.R, R/spark-model.R
\name{train_model.forecast_model}
\alias{train_model.forecast_model}
\alias{train_model}
\alias{train_model.spark_model_clustering}
\alias{train_model.spark_model}
\title{Train Model Generic Function}
\usage{
\method{train_model}{forecast_model}(mobj, data, measure, samples,
  save_submodels, execution_strategy, level, ...)

train_model(mobj, ...)

\method{train_model}{spark_model_clustering}(mobj, data, measure, samples,
  save_submodels, execution_strategy = "sequential", ...)

\method{train_model}{spark_model}(mobj, data, measure, samples,
  save_submodels, execution_strategy, seed, ...)
}
\arguments{
\item{mobj}{model object}

\item{data}{data to use for model training}

\item{measure}{measure object}

\item{samples}{samples object}

\item{save_submodels}{logical flag to save sub model fits. Default is TRUE}

\item{execution_strategy}{name of evaluation function to use for resolving a
future. Used for model training execution. default is sequential}

\item{...}{additional arguments to pass to model class train_model method}
}
\value{
updated modeler object
}
\description{
Train a model added to a modeler object.
}
\details{
Function fits model based on modeler samples, model pipeline, model method
and param grid.
}
